+++ brainconn/clustering/__init__.py
@@ -1,0 +1,22 @@
+"""
+Metrics which group nodes within graphs into clusters.
+"""
+from .clustering import (agreement, agreement_weighted,
+                         clustering_coef_bd, clustering_coef_bu,
+                         clustering_coef_wd, clustering_coef_wu,
+                         clustering_coef_wu_sign,
+                         consensus_und,
+                         get_components, get_components_old,
+                         number_of_components,
+                         transitivity_bd, transitivity_bu,
+                         transitivity_wd, transitivity_wu)
+
+__all__ = ['agreement', 'agreement_weighted',
+           'clustering_coef_bd', 'clustering_coef_bu',
+           'clustering_coef_wd', 'clustering_coef_wu',
+           'clustering_coef_wu_sign',
+           'consensus_und',
+           'get_components', 'get_components_old',
+           'number_of_components',
+           'transitivity_bd', 'transitivity_bu',
+           'transitivity_wd', 'transitivity_wu']