+++ brainconn/core/core.py
@@ -3,8 +3,9 @@ Miscellaneous functions.
 """
 from __future__ import division, print_function
 import numpy as np
-from .degree import degrees_dir, degrees_und, strengths_dir, strengths_und
-from .degree import strengths_und_sign
+from ..degree import degrees_dir, degrees_und, strengths_dir, strengths_und
+from ..degree import strengths_und_sign
+from ..utils import BCTParamError
 
 
 def assortativity_bin(CIJ, flag=0):
@@ -164,13 +165,13 @@ def core_periphery_dir(W, gamma=1, C0=None):
     n = len(W)
     np.fill_diagonal(W, 0)
 
-    if C0 == None:
+    if C0 is None:
         C = np.random.randint(2, size=(n,))
     else:
         C = C0.copy()
 
-    #methodological note, the core-detection null model is not corrected
-    #for degree cf community detection (to enable detection of hubs)
+    # methodological note, the core-detection null model is not corrected
+    # for degree cf community detection (to enable detection of hubs)
 
     s = np.sum(W)
     p = np.mean(W)
@@ -181,7 +182,7 @@ def core_periphery_dir(W, gamma=1, C0=None):
     q = np.sum(B[np.ix_(cix, cix)]) - np.sum(B[np.ix_(ncix, ncix)])
 
     print(q)
-    #sqish
+    # sqish
 
     flag = True
     it = 0
@@ -191,7 +192,7 @@ def core_periphery_dir(W, gamma=1, C0=None):
             raise BCTParamError('Infinite Loop aborted')
 
         flag = False
-        #initial node indices
+        # initial node indices
         ixes = np.arange(n)
 
         Ct = C.copy()
@@ -209,12 +210,12 @@ def core_periphery_dir(W, gamma=1, C0=None):
             print(np.where(np.abs(Qt[ixes]-max_Qt) < 1e-10))
             print(Qt[ixes])
             print(max_Qt)
-            #tunourn
+            # tunourn
             u = u[np.random.randint(len(u))]
             print(np.sum(Ct))
             Ct[ixes[u]] = np.logical_not(Ct[ixes[u]])
             print(np.sum(Ct))
-            #casga
+            # casga
 
             ixes = np.delete(ixes, u)
 
@@ -403,17 +404,20 @@ def local_assortativity_wu_sign(W):
     for curr_node in range(n):
         jp = np.where(W[curr_node, :] > 0)
         loc_assort_pos[curr_node] = np.sum(np.abs(str_pos[jp] -
-            str_pos[curr_node])) / str_pos[curr_node]
+                                                  str_pos[curr_node])) /\
+            str_pos[curr_node]
         jn = np.where(W[curr_node, :] < 0)
         loc_assort_neg[curr_node] = np.sum(np.abs(str_neg[jn] -
-            str_neg[curr_node])) / str_neg[curr_node]
+                                                  str_neg[curr_node])) /\
+            str_neg[curr_node]
 
-    loc_assort_pos = ((r_pos + 1) / n -
-        loc_assort_pos / np.sum(loc_assort_pos))
-    loc_assort_neg = ((r_neg + 1) / n -
-        loc_assort_neg / np.sum(loc_assort_neg))
+    loc_assort_pos = ((r_pos + 1) / n - loc_assort_pos /
+                      np.sum(loc_assort_pos))
+    loc_assort_neg = ((r_neg + 1) / n - loc_assort_neg /
+                      np.sum(loc_assort_neg))
 
     return loc_assort_pos, loc_assort_neg
+
 
 def rich_club_bd(CIJ, klevel=None):
     """
@@ -487,7 +491,7 @@ def rich_club_bu(CIJ, klevel=None):
     """
     deg = degrees_und(CIJ)  # compute degree of each node
 
-    if klevel == None:
+    if klevel is None:
         klevel = int(np.max(deg))
 
     R = np.zeros((klevel,))
@@ -521,7 +525,7 @@ def rich_club_wd(CIJ, klevel=None):
     Rw : Kx1 :obj:`numpy.ndarray`
         vector of rich-club coefficients for levels 1 to klevel
     """
-    nr_nodes = len(CIJ)
+    # nr_nodes = len(CIJ)
     # degree of each node is defined here as in+out
     deg = np.sum((CIJ != 0), axis=0) + np.sum((CIJ.T != 0), axis=0)
 
@@ -568,7 +572,7 @@ def rich_club_wu(CIJ, klevel=None):
     Rw : Kx1 :obj:`numpy.ndarray`
         vector of rich-club coefficients for levels 1 to klevel
     """
-    nr_nodes = len(CIJ)
+    # nr_nodes = len(CIJ)
     deg = np.sum((CIJ != 0), axis=0)
 
     if klevel is None: