+++ brainconn/modularity/modularity.py
@@ -131,14 +131,14 @@ def community_louvain(W, gamma=1, ci=None, B='modularity', seed=None):
         s1 = np.sum(W1)
         if s1:
             B1 = (W1 - gamma * np.outer(np.sum(W1, axis=1), np.sum(W1, axis=0))
-                / s1)
+                  / s1)
         else:
             B1 = 0
 
     elif np.min(W) < -1e-10:
-        raise BCTParamError("Input connection matrix contains negative "
-            'weights but objective function dealing with negative weights '
-            'was not selected')
+        raise BCTParamError('Input connection matrix contains negative '
+                            'weights but objective function dealing with '
+                            'negative weights was not selected')
 
     if B == 'potts' and np.any(np.logical_not(np.logical_or(W == 0, W == 1))):
         raise BCTParamError('Potts hamiltonian requires binary input matrix')
@@ -154,15 +154,15 @@ def community_louvain(W, gamma=1, ci=None, B='modularity', seed=None):
     else:
         try:
             B = np.array(B)
-        except:
-            raise BCTParamError('unknown objective function type')
+        except BCTParamError:
+            print('unknown objective function type')
 
         if B.shape != W.shape:
             raise BCTParamError('objective function matrix does not match '
                                 'size of adjacency matrix')
         if not np.allclose(B, B.T):
-            print ('Warning: objective function matrix not symmetric, '
-                   'symmetrizing')
+            print('Warning: objective function matrix not symmetric, '
+                  'symmetrizing')
             B = (B + B.T) / 2
 
     Hnm = np.zeros((n, n))
@@ -336,28 +336,26 @@ def link_communities(W, type_clustering='single'):
     U = np.arange(m)  # initial community assignments
     C[0, :] = np.arange(m)
 
-    import time
-
     for i in range(m - 1):
         print('hierarchy %i' % i)
 
-        #time1 = time.time()
+        # time1 = time.time()
 
         for j in range(len(U)):  # loop over communities
             ixes = C[i, :] == U[j]  # get link indices
 
             links = np.sort(Lw[ixes])
-            #nodes = np.sort(Ln[ixes,:].flat)
+            # nodes = np.sort(Ln[ixes,:].flat)
 
             nodes = np.sort(np.reshape(
                 Ln[ixes, :], 2 * np.size(np.where(ixes))))
 
             # get unique nodes
             nodulo = np.append(nodes[0], (nodes[1:])[nodes[1:] != nodes[:-1]])
-            #nodulo = ((nodes[1:])[nodes[1:] != nodes[:-1]])
+            # nodulo = ((nodes[1:])[nodes[1:] != nodes[:-1]])
 
             nc = len(nodulo)
-            #nc = len(nodulo)+1
+            # nc = len(nodulo)+1
             mc = np.sum(links)
             min_mc = np.sum(links[:nc - 1])  # minimal weight
             dc = (mc - min_mc) / (nc * (nc - 1) /
@@ -371,14 +369,14 @@ def link_communities(W, type_clustering='single'):
             Mc[i, j] = mc
             Dc[i, j] = dc if not np.isnan(dc) else 0
 
-        #time2 = time.time()
-        #print('compute densities time', time2-time1)
+        # time2 = time.time()
+        # print('compute densities time', time2-time1)
 
         C[i + 1, :] = C[i, :]  # copy current partition
 
-        #if i in (2693,):
-        #    import pdb
-        #    pdb.set_trace()
+        # if i in (2693,):
+        #     import pdb
+        #     pdb.set_trace()
 
         # Profiling and debugging show that this line, finding
         # the max values in this matrix, take about 3x longer than the
@@ -396,19 +394,19 @@ def link_communities(W, type_clustering='single'):
             u1 = uc
             u2 = ud
 
-        #time25 = time.time()
-        #print('copy and max time', time25-time2)
+        # time25 = time.time()
+        # print('copy and max time', time25-time2)
 
         # get unique links (implementation of matlab sortrows)
-        #ugl = np.array((u1,u2))
+        # ugl = np.array((u1,u2))
         ugl = np.sort((u1, u2), axis=1)
         ug_rows = ugl[np.argsort(ugl, axis=0)[:, 0]]
         # implementation of matlab unique(A, 'rows')
         unq_rows = np.vstack({tuple(row) for row in ug_rows})
         V = U[unq_rows]
 
-        #time3 = time.time()
-        #print('sortrows time', time3-time25)
+        # time3 = time.time()
+        # print('sortrows time', time3-time25)
 
         for j in range(len(V)):
             if type_clustering == 'single':
@@ -430,23 +428,17 @@ def link_communities(W, type_clustering='single'):
             C[i + 1, C[i + 1, :] == V[j, 1]] = V[j, 0]
             V[V == V[j, 1]] = V[j, 0]
 
-        #time4 = time.time()
-        #print('get linkages time', time4-time3)
+        # time4 = time.time()
+        # print('get linkages time', time4-time3)
 
         U = np.unique(C[i + 1, :])
         if len(U) == 1:
             break
 
-        #time5 = time.time()
-        #print('get unique communities time', time5-time4)
+        # time5 = time.time()
+        # print('get unique communities time', time5-time4)
 
-    #ENDT HAIERARKIKL CLUSTRRINNG
-    #ENDT HAIERARKIKL CLUSTRRINNG
-    #ENDT HAIERARKIKL CLUSTRRINNG
-    #ENDT HAIERARKIKL CLUSTRRINNG
-    #ENDT HAIERARKIKL CLUSTRRINNG
-
-    #Dc[ np.where(np.isnan(Dc)) ]=0
+    # Dc[ np.where(np.isnan(Dc)) ]=0
     i = np.argmax(np.sum(Dc * Mc, axis=1))
 
     U = np.unique(C[i, :])
@@ -529,10 +521,10 @@ def modularity_dir(A, gamma=1, kci=None):
                     (np.dot(modmat, mod_asgn_iter))
                 qmax = np.max(q_iter * it)
                 imax = np.argmax(q_iter * it)
-                #imax, = np.where(q_iter == qmax)
-                #if len(imax) > 0:
-                #    imax = imax[0]
-                #    print(imax)
+                # imax, = np.where(q_iter == qmax)
+                # if len(imax) > 0:
+                #     imax = imax[0]
+                #     print(imax)
                 # does switching increase modularity?
                 mod_asgn_iter[imax] *= -1
                 it[imax] = np.ma.masked
@@ -699,7 +691,7 @@ def modularity_finetune_und(W, ci=None, gamma=1, seed=None):
     """
     np.random.seed(seed)
 
-    #import time
+    # import time
     n = len(W)  # number of nodes
     if ci is None:
         ci = np.arange(n) + 1
@@ -959,7 +951,7 @@ def modularity_louvain_dir(W, gamma=1, hierarchy=False, seed=None):
         while flag:
             it += 1
             if it > 1000:
-                raise BCTParamError('Modularity Infinite Loop Style F.  Please '
+                raise BCTParamError('Modularity Infinite Loop Style F. Please '
                                     'contact the developer with this error.')
             flag = False
 
@@ -1070,7 +1062,7 @@ def modularity_louvain_und(W, gamma=1, hierarchy=False, seed=None):
     q.append(-1)  # hierarchical modularity values
     n0 = n
 
-    #knm = np.zeros((n,n))
+    # knm = np.zeros((n,n))
     # for j in np.xrange(n0+1):
     #    knm[:,j] = np.sum(w[;,
 
@@ -1089,7 +1081,7 @@ def modularity_louvain_und(W, gamma=1, hierarchy=False, seed=None):
         while flag:
             it += 1
             if it > 1000:
-                raise BCTParamError('Modularity Infinite Loop Style C.  Please '
+                raise BCTParamError('Modularity Infinite Loop Style C. Please '
                                     'contact the developer with this error.')
             flag = False
 
@@ -1249,15 +1241,18 @@ def modularity_louvain_und_sign(W, gamma=1, qtype='sta', seed=None):
             it += 1
             if it > 1000:
                 raise BCTParamError('Infinite Loop was detected and stopped. '
-                                    'This was probably caused by passing in a directed matrix.')
+                                    'This was probably caused by passing in a '
+                                    'directed matrix.')
             flag = False
             # loop over nodes in random order
             for u in np.random.permutation(nh):
                 ma = m[u] - 1
+                # positive dQ
                 dQ0 = ((knm0[u, :] + W0[u, u] - knm0[u, ma]) -
-                       gamma * kn0[u] * (km0 + kn0[u] - km0[ma]) / s0)  # positive dQ
+                       gamma * kn0[u] * (km0 + kn0[u] - km0[ma]) / s0)
+                # negative dQ
                 dQ1 = ((knm1[u, :] + W1[u, u] - knm1[u, ma]) -
-                       gamma * kn1[u] * (km1 + kn1[u] - km1[ma]) / s1)  # negative dQ
+                       gamma * kn1[u] * (km1 + kn1[u] - km1[ma]) / s1)
 
                 dQ = d0 * dQ0 - d1 * dQ1  # rescaled changes in modularity
                 dQ[ma] = 0  # no changes for same module
@@ -1518,9 +1513,9 @@ def modularity_und(A, gamma=1, kci=None):
                     (np.dot(modmat, mod_asgn_iter))
                 qmax = np.max(q_iter * it)
                 imax = np.argmax(q_iter * it)
-                #imax, = np.where(q_iter == qmax)
-                #if len(imax) > 1:
-                #    imax = imax[0]
+                # imax, = np.where(q_iter == qmax)
+                # if len(imax) > 1:
+                #     imax = imax[0]
                 # does switching increase modularity?
                 mod_asgn_iter[imax] *= -1
                 it[imax] = np.ma.masked
@@ -1594,8 +1589,8 @@ def modularity_und_sign(W, ci, qtype='sta'):
 
     Kn0 = np.sum(Knm0, axis=1)  # positive node degree
     Kn1 = np.sum(Knm1, axis=1)  # negative node degree
-    Km0 = np.sum(Knm0, axis=0)  # positive module degree
-    Km1 = np.sum(Knm1, axis=0)  # negaitve module degree
+    # Km0 = np.sum(Knm0, axis=0)  # positive module degree
+    # Km1 = np.sum(Knm1, axis=0)  # negaitve module degree
 
     if qtype == 'smp':
         d0 = 1 / s0
@@ -1670,9 +1665,9 @@ def partition_distance(cx, cy):
     Py = np.histogram(cy, bins=np.max(cy))[0] / n
     Pxy = np.histogram(cxy, bins=np.max(cxy))[0] / n
 
-    Hx = -np.sum(Px * np.log(Px))
-    Hy = -np.sum(Py * np.log(Py))
-    Hxy = -np.sum(Pxy * np.log(Pxy))
+    Hx = -1 * np.sum(Px * np.log(Px))
+    Hy = -1 * np.sum(Py * np.log(Py))
+    Hxy = -1 * np.sum(Pxy * np.log(Pxy))
 
     Vin = (2 * Hxy - Hx - Hy) / np.log(n)
     Min = 2 * (Hx + Hy - Hxy) / (Hx + Hy)