+++ brainconn/tests/test_centrality.py
@@ -1,3 +1,5 @@
+# emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4 -*-
+# ex: set sts=4 ts=4 sw=4 et:
 import os.path as op
 
 import numpy as np
@@ -12,7 +14,7 @@ def test_pc():
 
     pc = np.load(op.join(MAT_DIR, 'sample_pc.npy'))
 
-    pc_ = bc.participation_coef(x, ci)
+    pc_ = bc.centrality.participation_coef(x, ci)
     print(list(zip(pc, pc_)))
 
     assert np.allclose(pc, pc_, atol=0.02)
@@ -22,42 +24,46 @@ def participation_test():
     W = np.eye(3)
     ci = np.array([1, 1, 2])
 
-    assert np.allclose(bc.participation_coef(W, ci), [0, 0, 0])
-    assert np.allclose(bc.participation_coef_sign(W, ci)[0], [0, 0, 0])
+    assert np.allclose(bc.centrality.participation_coef(W, ci), [0, 0, 0])
+    assert np.allclose(bc.centrality.participation_coef_sign(W, ci)[0],
+                       [0, 0, 0])
 
     W = np.ones((3, 3))
-    assert np.allclose(bc.participation_coef(W, ci), [
-                       0.44444444, 0.44444444, 0.44444444])
-    assert np.allclose(bc.participation_coef_sign(W, ci)[0], [
-                       0.44444444, 0.44444444, 0.44444444])
+    assert np.allclose(bc.centrality.participation_coef(W, ci), [
+        0.44444444, 0.44444444, 0.44444444])
+    assert np.allclose(bc.centrality.participation_coef_sign(W, ci)[0], [
+        0.44444444, 0.44444444, 0.44444444])
 
     W = np.eye(3)
     W[0, 1] = 1
     W[0, 2] = 1
-    assert np.allclose(bc.participation_coef(W, ci), [0.44444444, 0, 0])
-    assert np.allclose(bc.participation_coef_sign(W, ci)
-                       [0], [0.44444444, 0, 0])
+    assert np.allclose(bc.centrality.participation_coef(W, ci),
+                       [0.44444444, 0, 0])
+    assert np.allclose(bc.centrality.participation_coef_sign(W, ci)[0],
+                       [0.44444444, 0, 0])
 
     W = np.eye(3)
     W[0, 1] = -1
     W[0, 2] = -1
     W[1, 2] = 1
-    assert np.allclose(bc.participation_coef_sign(W, ci)[0], [0.,  0.5,  0.])
+    assert np.allclose(bc.centrality.participation_coef_sign(W, ci)[0],
+                       [0.,  0.5,  0.])
 
 
 def gateway_test():
     x = load_sample(thres=.1)
     ci = np.load(op.join(MAT_DIR, 'sample_partition.npy'))
 
-    g_pos, _ = bc.gateway_coef_sign(x, ci)
+    g_pos, _ = bc.centrality.gateway_coef_sign(x, ci)
 
     print(np.sum(g_pos), 43.4382)
     assert np.allclose(np.sum(g_pos), 43.4382, atol=.001)
 
-    g_pos_bet, _ = bc.gateway_coef_sign(x, ci, centrality_type='betweenness')
+    gpb, _ = bc.centrality.gateway_coef_sign(x, ci,
+                                             centrality_type='betweenness')
 
-    print(np.sum(g_pos_bet), 43.4026)
-    assert np.allclose(np.sum(g_pos_bet), 43.4026, atol=.001)
+    print(np.sum(gpb), 43.4026)
+    assert np.allclose(np.sum(gpb), 43.4026, atol=.001)
 
 
 def test_zi():
@@ -66,7 +72,7 @@ def test_zi():
 
     zi = np.load(op.join(MAT_DIR, 'sample_zi.npy'))
 
-    zi_ = bc.module_degree_zscore(x, ci)
+    zi_ = bc.centrality.module_degree_zscore(x, ci)
     print(list(zip(zi, zi_)))
 
     assert np.allclose(zi, zi_, atol=0.05)
@@ -82,8 +88,7 @@ def test_zi():
 def test_shannon_entropy():
     x = load_sample(thres=0.4)
     ci = np.load(op.join(MAT_DIR, 'sample_partition.npy'))
-    # ci, q = bc.modularity_und(x)
-    hpos, _ = bc.diversity_coef_sign(x, ci)
+    hpos, _ = bc.centrality.diversity_coef_sign(x, ci)
     print(np.sum(hpos))
     print(hpos[-1])
     assert np.allclose(np.sum(hpos), 102.6402, atol=.01)